Add dry flag

+++ src/main.rs
@@ -10,8 +10,12 @@ use std::sync::{atomic, RwLock};
 use structopt::StructOpt;
 
 #[derive(Debug, StructOpt)]
-#[structopt(name = "cleanup", about = "Recursively cleanup a directory by removing duplicate files")]
+#[structopt(name = "cleanup", about = "Recurisvely find and remove duplicate files in a target directory")]
 struct Opt {
+    /// Do a dry run
+    #[structopt(long = "dry")]
+    dry: bool,
+
     /// Input directory
     #[structopt(parse(from_os_str))]
     input: PathBuf,
@@ -21,16 +25,24 @@ fn main() -> Result<()> {
     let opt = Opt::from_args();
     let deleted_count = atomic::AtomicUsize::new(0);
 
+    let input = opt.input.clone();
     rayon::scope(|s| {
         s.spawn(|s| {
-            process_directory(opt.input, &deleted_count, s);
+            process_directory(input, &deleted_count, opt.dry, s);
         });
     });
 
-    println!(
-        "Deleted {} files",
-        deleted_count.load(atomic::Ordering::Relaxed)
-    );
+    if opt.dry {
+        println!(
+            "Would have deleted {} files",
+            deleted_count.load(atomic::Ordering::Relaxed)
+        );
+    } else {
+        println!(
+            "Deleted {} files",
+            deleted_count.load(atomic::Ordering::Relaxed)
+        );
+    }
 
     Ok(())
 }
@@ -40,9 +52,9 @@ fn main() -> Result<()> {
 fn process_directory<'a, 'b>(
     dir: PathBuf,
     deleted_file_count: &'a AtomicUsize,
+    is_dry_run: bool,
     scope: &'b Scope<'a>,
 ) {
-    println!("{:?}", dir);
     let hashes = RwLock::new(HashSet::new());
     for entry in fs::read_dir(dir).expect("failed to read dir") {
         let mut hasher = Sha1::new();
@@ -52,7 +64,7 @@ fn process_directory<'a, 'b>(
         let file_type = entry.file_type().expect("failed to get file type");
         if file_type.is_dir() && !file_type.is_symlink() {
             scope.spawn(move |s| {
-                process_directory(path, deleted_file_count, s);
+                process_directory(path, deleted_file_count, is_dry_run, s);
             });
             continue;
         }
@@ -65,7 +77,10 @@ fn process_directory<'a, 'b>(
         let result = hasher.finalize();
         if hashes.read().unwrap().contains(&result) {
             deleted_file_count.fetch_add(1, atomic::Ordering::Relaxed);
-            fs::remove_file(&path).expect("failed to remove file");
+
+            if !is_dry_run {
+                fs::remove_file(&path).expect("failed to remove file");
+            }
         } else {
             hashes.write().unwrap().insert(result);
         }
